Added alternate "instrument extraction" for simultaneous model evaluation
Added new method of evaluating simultaneous model evaluations. Instead of interpolating
over a dummy grid (which can introduce artifacts) or requiring the user to add their own
grids, we instead concatenate the relevant grids together, find all unique instances of
values (energy, frequency, etc) and create extraction masks of the grid for each data-product.
When the model is evaluated, we iterate through the different masks, applying them to the single
model result, retrieving only the values related to that object. The results are then
convolved with the instrument response when appropriate, and then collated into the
stack of results as previously.